Refactor map rendering assets
